% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energygof.R
\name{egof}
\alias{egof}
\title{Energy goodness-of-fit tests for univariate distributions}
\usage{
egof(
  x,
  dist = c("uniform", "exponential", "bernoulli", "binomial", "geometric", "normal",
    "gaussian", "beta", "poisson", "lognormal", "lnorm", "laplace", "doubleexponential",
    "asymmetriclaplace", "inversegaussian", "standardhalfnormal", "halfnormal", "chisq",
    "chisquared", "gamma", "weibull", "cauchy", "pareto"),
  R = 100,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{dist}{A string. The distribution to test.}

\item{R}{A positive integer. The number of parametric bootstrap replicates
taken to calculate the p-value.}

\item{...}{Parameters of the distribution \code{dist}. For distributions in
the R `stats' library, parameter argument names are identical. To test the
_composite_ goodness-of-fit hypothesis that \code{x} is distributed
according to the _family of distributions_ \code{dist}, don't pass
parameters in \code{...}. #'}
}
\value{
An object of class `htest' representing the result of the energy
  goodness-of-fit hypothesis test. The htest object has the elements:

\describe{
  \item{\code{method}}{A test description.}
  \item{\code{data.name}}{Name of data passed in x.}
  \item{\code{parameters}}{What was passed in ..., or NULL for composite tests.}
  \item{\code{null-value}}{A description of the null hypothesis.}
  \item{\code{R}}{Number of simulation replicates.}
  \item{\code{composite_p}}{A logical, TRUE if composite test was performed.}
  \item{\code{statistic}}{NULL, or a list of MLEs calculated, if a composite test was performed.}
  \item{\code{p.value}}{A numeric p-value.}
  \item{\code{estimate}}{A numeric value of the energy statistic for testing \code{x} against {dist}.}
}
}
\description{
Energy goodness-of-fit tests for univariate distributions
}
\details{
[TODO description of Energy GOF test here.]

There are two types of goodness-of-fit tests covered by the egof function,
simple and composite. Simple GOF tests test the data `x` against a specific
distribution with _known parameters_ that you must pass to egof in the
ellipsis agrument (...). You should use a simple GOF test if you wish to
test questions like "my data is Normal with mean 1 and sd 2". egof can also
conduct _some_ composite GOF tests. A composite test is performed if no
parameters are passed in the ellpisis argument (...). You should conduct a
composite test if your research question is "my data is Normal."

All the composite tests in egof assume that none of the parameters are
known. So while there is a statistical test of Normality with known mean and
unknown sd, this is not implemented in the energygof package. So, either
pass all the distribution parameters or none of them. (In the special case
of the Normal distribution, you can use the energy package to test the GOF
hypothesis with any combination of known and known parameters.)

You should set R to be a very large number in practice. I recommend at least
10,000. The default value is not a robust choice.
}
\examples{
x <- rnorm(10)

## Composite energy goodness-of-fit test (test for Normality with unknown
## parameters)

egof(x, "normal", R = 10)

## Simple energy goodness-of-fit test (test for Normality with known
## parameters)

egof(x, "normal", R = 10, mean = 0, sd = 1)

## Simple energy goodness-of-fit test for Weibull distribution

y <- rweibull(10, 1, 1)
egof(y, "weibull", shape = 1, scale = 3)

## Error, egof does not support "partially composite" GOF tests

## Not run:
egof(x, "normal", R = 10, mean = 0)
## End(Not run)

}
\references{
Székely, G. J., & Rizzo, M. L. (2023). The energy of data and distance
correlation. Chapman and Hall/CRC.

Székely, G. J., & Rizzo, M. L. (2013). Energy statistics: A class of
statistics based on distances. Journal of statistical planning and
inference, 143(8), 1249-1272.

Li, Y. (2015). Goodness-of-fit tests for Dirichlet distributions with
applications. Bowling Green State University.

Rizzo, M. L. (2002). A new rotation invariant goodness-of-fit test (PhD
thesis). Bowling Green State University

Haman, J. T. (2018). The energy goodness-of-fit test and EM type estimator
for asymmetric Laplace distributions (Doctoral dissertation, Bowling Green
State University).

Ofosuhene, P. (2020). The energy goodness-of-fit test for the inverse
Gaussian distribution (Doctoral dissertation, Bowling Green State
University).

Rizzo, M. L. (2009). New goodness-of-fit tests for Pareto distributions.
ASTIN Bulletin: The Journal of the IAA, 39(2), 691-715.
}
\seealso{
\link[stats]{Distributions} for a list of distributions available
  in most R installations. \link[energy]{normal.test} for the energy
  goodness-of-fit test with unknown parameters. \link[energy]{normal.e} for
  the energy goodness-of-fit statistic. See the
  \link[energy]{poission.mtest} for a different poisson goodness-of-fit test
  based on mean distances. The tests for Normal and Poisson distribution in
  the \link[energy] package are implemented in C/C++ , and are faster than
  the ones available in the egof package.
}
\author{
John T. Haman
}
