% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{energyfit}
\alias{energyfit}
\alias{ef}
\title{Goodness-of-fit tests via the energy of data}
\usage{
energyfit(x, dist, nsim = 100)
}
\arguments{
\item{x}{A numeric vector.}

\item{dist}{A distribution object. The distribution to test \code{x} against.
Distribution objects are created with the various \code{name_dist(...)}
functions in this package.}

\item{nsim}{A positive integer. The number of parametric bootstrap replicates
taken to calculate the p-value.}
}
\value{
An object of class `htest' representing the result of the energy
goodness-of-fit hypothesis test. The htest object has the list elements:
\itemize{
\item method: Simple or Composite
\item data.name
\item distribution: The distribution object created to test
\item parameter: List of parameters if the test is simple
\item nsim: Number of bootstrap replicates
\item pow: If generalized energy test, the exponent of the test
\item composite_p: TRUE/FALSE composite predicate
\item statistic: The value of the energy statistic (see \link{Qhat})
\item p.value
\item sim_reps: bootstrap simulation data
\item estimate: Any parameter estimates, if the test is composite
}
}
\description{
This is an alternative interface that provides the same test as
ef.test, but allows the user to pass a distribution object. The advantage
is that you do not need to pass distribution parameters into a \code{...}
argument as in \code{ef.test}. \code{ef.test} uses this function under the hood, but
it's perfectly suitable for the user to use as well.
}
\details{
\link{TODO description of Energy GOF test here.}

There are two types of goodness-of-fit tests covered by the energyfit function,
simple and composite. Simple GOF tests test the data \code{x} against a specific
distribution with \emph{known parameters} that you must pass to energyfit in the
ellipsis agrument (...). You should use a simple GOF test if you wish to
test questions like "my data is Normal with mean 1 and sd 2". energyfit can also
conduct \emph{some} composite GOF tests. A composite test is performed if no
parameters are passed in the ellpisis argument (...). You should conduct a
composite test if your research question is "my data is Normal."

All the composite tests in energyfit assume that none of the parameters are
known. So while there is a statistical test of Normality with known mean and
unknown sd, this is not implemented in the energygof package. So, either
pass all the distribution parameters or none of them. (In the special case
of the Normal distribution, you can use the energy package to test the GOF
hypothesis with any combination of known and known parameters.)

You should set nsim to be a very large number in practice. I recommend at least
10,000. The default value is not a robust choice.
}
\examples{

ef(rnorm(10), normal_dist(0,1), nsim = 10)

ef(rpois(10,1), poisson_dist(1)) # No p-value

ef(rnorm(10), normal_dist(), nsim = 10) # Composite test

[TODO] example with quantile function.

}
\references{
Székely, G. J., & Rizzo, M. L. (2023). The energy of data and distance
correlation. Chapman and Hall/CRC.

Székely, G. J., & Rizzo, M. L. (2013). Energy statistics: A class of
statistics based on distances. Journal of statistical planning and
inference, 143(8), 1249-1272.

Li, Y. (2015). Goodness-of-fit tests for Dirichlet distributions with
applications. Bowling Green State University.

Rizzo, M. L. (2002). A new rotation invariant goodness-of-fit test (PhD
thesis). Bowling Green State University

Haman, J. T. (2018). The energy goodness-of-fit test and EM type estimator
for asymmetric Laplace distributions (Doctoral dissertation, Bowling Green
State University).

Ofosuhene, P. (2020). The energy goodness-of-fit test for the inverse
Gaussian distribution (Doctoral dissertation, Bowling Green State
University).

Rizzo, M. L. (2009). New goodness-of-fit tests for Pareto distributions.
ASTIN Bulletin: The Journal of the IAA, 39(2), 691-715.
}
\seealso{
\link[stats]{Distributions} for a list of distributions available
in most R installations. \link[energy:normalGOF]{energy::normal.test} for the energy
goodness-of-fit test with unknown parameters. See
\link[energy:poisson]{energy::poisson.mtest} for a different poisson goodness-of-fit test
based on mean distances. The tests for (multivariate) Normal in the
\link{energy} package are implemented with compiled code, and are faster
than the one available in the energyfit package.
}
\author{
John T. Haman
}
