% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{energyfit.test}
\alias{energyfit.test}
\alias{ef.test}
\title{Goodness-of-fit tests via the energy of data}
\usage{
\method{energyfit}{test}(
  x,
  dist = c("uniform", "exponential", "bernoulli", "binomial", "geometric", "normal",
    "gaussian", "beta", "poisson", "lognormal", "lnorm", "laplace", "doubleexponential",
    "asymmetriclaplace", "alaplace", "inversegaussian", "halfnormal", "chisq",
    "chisquared", "gamma", "weibull", "cauchy", "stable", "pareto"),
  nsim = 100,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{dist}{A string. The distribution to test \code{x} against.}

\item{nsim}{A positive integer. The number of parametric bootstrap replicates
taken to calculate the p-value.}

\item{...}{Parameters of the distribution \code{dist}. For distributions in
the R \verb{stats' library, parameter argument names are identical. To test the _composite_ goodness-of-fit hypothesis that }x\verb{is distributed according to the _family of distributions_}dist\verb{, don't pass parameters in }...`}
}
\value{
An object of class `htest' representing the result of the energy
goodness-of-fit hypothesis test. The htest object has the list elements:
\itemize{
\item method: Simple or Composite
\item data.name
\item distribution: The distribution object created to test
\item parameter: List of parameters if the test is simple
\item nsim: Number of bootstrap replicates
\item pow: If generalized energy test, the exponent of the test
\item composite_p: TRUE/FALSE composite predicate
\item statistic: The value of the energy statistic (see \link{Qhat})
\item p.value
\item sim_reps: bootstrap simulation data
\item estimate: Any parameter estimates, if the test is composite
}
}
\description{
Goodness-of-fit tests via the energy of data
}
\details{
\link{TODO description of Energy GOF test here.}

There are two types of goodness-of-fit tests covered by the energyfit function,
simple and composite. Simple GOF tests test the data \code{x} against a specific
distribution with \emph{known parameters} that you must pass to energyfit in the
ellipsis agrument (...). You should use a simple GOF test if you wish to
test questions like "my data is Normal with mean 1 and sd 2". energyfit can also
conduct \emph{some} composite GOF tests. A composite test is performed if no
parameters are passed in the ellpisis argument (...). You should conduct a
composite test if your research question is "my data is Normal."

All the composite tests in energyfit assume that none of the parameters are
known. So while there is a statistical test of Normality with known mean and
unknown sd, this is not implemented in the energygof package. So, either
pass all the distribution parameters or none of them. (In the special case
of the Normal distribution, you can use the energy package to test the GOF
hypothesis with any combination of known and known parameters.)

You should set nsim to be a very large number in practice. I recommend at least
10,000. The default value is not a robust choice.
}
\examples{
x <- rnorm(10)

## Composite energy goodness-of-fit test (test for Normality with unknown
## parameters)

energyfit.test(x, "normal", nsim = 10)

## Simple energy goodness-of-fit test (test for Normality with known
## parameters). ef is an alias for energyfit.test.

ef.test(x, "normal", nsim = 10, mean = 0, sd = 1)

## Alternatively, use the energyfit generic directly so that you do not need
## to pass parameter names into `...`

energyfit(x, normal_dist(0, 1), nsim = 10)

## Simple energy goodness-of-fit test for Weibull distribution

y <- rweibull(10, 1, 1)
energyfit.test(y, "weibull", shape = 1, scale = 3, nsim = 10)

## Alternatively, use the energyfit generic directly, which is slightly less
## verbose. ef is an alias for energyfit.

ef(y, weibull_dist(1, 3), nsim = 10

## energyfit does not support "partially composite" GOF tests, so this will
## result in an error.

\dontrun{
energyfit.test(x, "normal", mean = 0, nsim = 10) # sd is missing
}

}
\references{
Székely, G. J., & Rizzo, M. L. (2023). The energy of data and distance
correlation. Chapman and Hall/CRC.

Székely, G. J., & Rizzo, M. L. (2013). Energy statistics: A class of
statistics based on distances. Journal of statistical planning and
inference, 143(8), 1249-1272.

Li, Y. (2015). Goodness-of-fit tests for Dirichlet distributions with
applications. Bowling Green State University.

Rizzo, M. L. (2002). A new rotation invariant goodness-of-fit test (PhD
thesis). Bowling Green State University

Haman, J. T. (2018). The energy goodness-of-fit test and EM type estimator
for asymmetric Laplace distributions (Doctoral dissertation, Bowling Green
State University).

Ofosuhene, P. (2020). The energy goodness-of-fit test for the inverse
Gaussian distribution (Doctoral dissertation, Bowling Green State
University).

Rizzo, M. L. (2009). New goodness-of-fit tests for Pareto distributions.
ASTIN Bulletin: The Journal of the IAA, 39(2), 691-715.
}
\seealso{
\link[stats]{Distributions} for a list of distributions available
in most R installations. \link[energy:normalGOF]{energy::normal.test} for the energy
goodness-of-fit test with unknown parameters. See
\link[energy:poisson]{energy::poisson.mtest} for a different poisson goodness-of-fit test
based on mean distances. The tests for (multivariate) Normal in the
\link{energy} package are implemented with compiled code, and are faster
than the one available in the energyfit package.
}
\author{
John T. Haman
}
