% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{EXXhat}
\alias{EXXhat}
\title{Calculate expected distance from sample}
\usage{
EXXhat(x, dist)
}
\arguments{
\item{x}{Numeric vector.}

\item{dist}{Distibution object (created by `"name"_dist(...)`)}
}
\value{
\eqn{\frac{1}{n^2} \sum_i \sum_j  |x_i - x_j|^s}, where s=`pow` is `dist$pow`.
}
\description{
Calculate expected distance from sample
}
\author{
John T. Haman
}
