% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyGOF.R
\name{f_dist}
\alias{f_dist}
\title{Create a F distribution object for energy testing}
\usage{
f_dist(df1 = NULL, df2 = NULL)
}
\arguments{
\item{df1}{Same as the shape parameter in \code{\link[stats:Fdist]{stats::rf()}}}

\item{df2}{Same as the scale parameter in \code{\link[stats:Fdist]{stats::rf()}}}
}
\value{
S3 data object containing following the fields.
\itemize{
\item \code{name}: String
\item \code{composite_p}: Composite predicate. TRUE if test is composite.
\item \code{par}: Distribution parameters, list of the formals.
\item \code{sampler_par}: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item \code{par_domain}: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item \code{support}: Function to check that data \code{x} can be tested against \code{dist}
\item \code{sampler}: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item \code{EYY}: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item \code{EXYhat}: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in \code{eg.test} or \code{eg}).
\item \code{xform}: Function that may be used to transform x. Only available in certain
distribution objects.
\item \code{statistic}: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item \code{notes}: Distribution specific messages. Only used in certain distribution
objects.
}
}
\description{
Create an S3 object that sets all the required data needed by
energyGOF to execute the energy goodness-of-fit test against a normal
distribution. If \code{mean} and \code{sd} are both NULL, perform a composite test.
}
\examples{

d <- f_dist(3, 3)
eg(rf(10, 3, 3), d)

}
\author{
John T. Haman
}
