% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{gamma_dist}
\alias{gamma_dist}
\title{Create a gamma distribution object for energy testing}
\usage{
gamma_dist(shape = NULL, rate = NULL)
}
\arguments{
\item{shape}{NULL, or same shape parameter in \code{\link[=rgamma]{rgamma()}} (must be length 1)}

\item{rate}{NULL, or same rate parameter in \code{\link[=rgamma]{rgamma()}} (must be length 1)}
}
\value{
S3 data object containing the fields. This all used internally by
\code{\link[=energyfit]{energyfit()}} and \code{\link[=energyfit.test]{energyfit.test()}}.
\itemize{
\item name: String
\item composite_p: TRUE if test is composite
\item par: Distribution parameters, list of the formals.
\item sampler_par: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item par_domain: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item support: Function to check that data \code{x} can be tested against \code{dist}
\item sampler: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item EYY: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item EXYhat: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in ef.test or ef).
\item xform: Function that may be used to transform x. Only available in certain
distribution objects.
\item statistic: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item notes: Distribution specific messages.
}
}
\description{
If both parameters are set to NULL, the composite test is
performed. In the composite case, the data are transformed to a chi-square
distribution, conditional on the maximum likelihood estimates. This is not a
complete pivot in the sense that the parameters are eliminated from the
distribution, but it's sort of close. A power analysis studies the operating
characteristics of this strategy in the vignette.
}
\author{
John T. Haman
}
