% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{inverse_gaussian_dist}
\alias{inverse_gaussian_dist}
\title{Create a inverse Gaussian distribution object for energy testing}
\usage{
inverse_gaussian_dist(mu = NULL, lambda = NULL)
}
\arguments{
\item{mu}{NULL or a positive mean parameter}

\item{lambda}{NULL or a positive shape parameter}
}
\value{
S3 data object containing the fields. This all used internally by
\code{\link[=energyfit]{energyfit()}} and \code{\link[=energyfit.test]{energyfit.test()}}.
\itemize{
\item name: String
\item composite_p: TRUE if test is composite
\item par: Distribution parameters, list of the formals.
\item sampler_par: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item par_domain: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item support: Function to check that data \code{x} can be tested against \code{dist}
\item sampler: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item EYY: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item EXYhat: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in ef.test or ef).
\item xform: Function that may be used to transform x. Only available in certain
distribution objects.
\item statistic: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item notes: Distribution specific messages.
}
}
\description{
This is exactly the distribution corresponding to the pdf

\deqn{
  f(x | \mu, \lambda) =
  \left( \frac{\lambda}{2 \pi x^3} \right)^{1/2}
  \exp \left( -\frac{\lambda (x - \mu)^2}{2 \mu^2 x} \right),
  \qquad x > 0.
}

If mu and lambda are both NULL, a composite test is performed.
}
\author{
John T. Haman
}
