% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyGOF.R
\name{laplace_dist}
\alias{laplace_dist}
\title{Create a Laplace distribution object for energy testing}
\usage{
laplace_dist(location = NULL, scale = NULL)
}
\arguments{
\item{location}{NULL, or the median of the distribution}

\item{scale}{NULL or a positive scale parameter}
}
\value{
S3 data object containing following the fields.
\itemize{
\item name: String
\item composite_p: Composite predicate. TRUE if test is composite.
\item par: Distribution parameters, list of the formals.
\item sampler_par: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item par_domain: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item support: Function to check that data \code{x} can be tested against \code{dist}
\item sampler: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item EYY: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item EXYhat: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in \code{eg.test} or \code{eg}).
\item xform: Function that may be used to transform x. Only available in certain
distribution objects.
\item statistic: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item notes: Distribution specific messages. Only used in certain distribution
objects.
}
}
\description{
Create an S3 object that sets all the required data needed by
energyGOF to execute the energy goodness-of-fit test against a Laplace
distribution. If location and scale are both NULL, a composite test is
performed.

This is exactly the distribution corresponding to the PDF

\deqn{f(x|\mu, b) = \frac{1}{2b} \exp \{ - \frac{|x - \mu}{b}\}, }

where \code{location} = \eqn{mu} and \code{scale} = \eqn{b}.
}
\examples{
d <- laplace_dist(1, 1)

x <- d$sampler(10, d$par)

eg(x, d)

}
\author{
John T. Haman
}
