% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyGOF.R
\name{normal_dist}
\alias{normal_dist}
\title{Create a Normal distribution object for energy testing}
\usage{
normal_dist(mean = NULL, sd = NULL)
}
\arguments{
\item{mean}{NULL, or if specified, same as \code{\link[=rnorm]{rnorm()}}, but must be length 1.}

\item{sd}{NULL, or if specified, Same as \code{\link[=rnorm]{rnorm()}}, but must be length 1}
}
\value{
S3 data object containing following the fields.
\itemize{
\item \code{name}: String
\item \code{composite_p}: Composite predicate. TRUE if test is composite.
\item \code{par}: Distribution parameters, list of the formals.
\item \code{sampler_par}: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item \code{par_domain}: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item \code{support}: Function to check that data \code{x} can be tested against \code{dist}
\item \code{sampler}: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item \code{EYY}: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item \code{EXYhat}: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in \code{eg.test} or \code{eg}).
\item \code{xform}: Function that may be used to transform x. Only available in certain
distribution objects.
\item \code{statistic}: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item \code{notes}: Distribution specific messages. Only used in certain distribution
objects.
}
}
\description{
Create an S3 object that sets all the required data needed by
energyGOF to execute the energy goodness-of-fit test against a normal
distribution. If \code{mean} and \code{sd} are both NULL, perform a composite test.
}
\examples{
d <- normal_dist(0, 1)

# Composite test
dc <- normal_dist()
eg(rnorm(10), dc)


### Expected distances:

d$EYY(d$par)

## should be about the same as mean(abs(rnorm(1e5) - rnorm(1e5)))

x <- 3

d$EXYhat(3, d$par)

## should be about the same as mean(abs(x - rnorm(1e5)))

}
\author{
John T. Haman
}
