% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{Qhat.CauchyDist}
\alias{Qhat.CauchyDist}
\title{Calculate Sample Energy Goodness-of-Fit Statistic}
\usage{
\method{Qhat}{CauchyDist}(x, dist, EYY)
}
\arguments{
\item{x}{Numeric vector.}

\item{dist}{Distibution object (created by `"name"_dist(...)`)}

\item{EYY}{The value of \eqn{E|Y-Y'|} (which is calculated with `dist$EYY()`, but passed as a separate argument because it's computation may be costly)}
}
\value{
Energy goodness-of-fit statistic for testing that `x` follows the distribution `dist`. In the generalized form, this is \deqn{Q = nE, where, }

\deqn{E = \frac{2}{n} E|x_i - Y|^s - E|Y-Y'|^s - \frac{1}{n^2} \sum_i \sum_j |x_i - x_j|^s}
}
\description{
Calculate Sample Energy Goodness-of-Fit Statistic
}
\author{
John T. Haman
}
