% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{stable_dist}
\alias{stable_dist}
\title{Create a stable distribution object for energy testing}
\usage{
stable_dist(
  location = 0,
  scale = 1,
  skew = 0,
  stability = 1,
  pow = stability/4
)
}
\arguments{
\item{location}{Same as in \code{\link[=rcauchy]{rcauchy()}}}

\item{scale}{Same as in \code{\link[=rcauchy]{rcauchy()}}}

\item{skew}{-1 < skew < 1 is required}

\item{stability}{The tail index or stability index. Controls the fatness of the tails. 0 < stability <= 2 is required.}

\item{pow}{Exponent of the energy test. 0 < stability/2 is required.}
}
\value{
S3 data object containing the fields. This all used internally by
\code{\link[=energyfit]{energyfit()}} and \code{\link[=energyfit.test]{energyfit.test()}}.
\itemize{
\item name: String
\item composite_p: TRUE if test is composite
\item par: Distribution parameters, list of the formals.
\item sampler_par: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item par_domain: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item support: Function to check that data \code{x} can be tested against \code{dist}
\item sampler: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item EYY: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item EXYhat: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in ef.test or ef).
\item xform: Function that may be used to transform x. Only available in certain
distribution objects.
\item statistic: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item notes: Distribution specific messages.
}
}
\description{
This is a very slow test due to an onerous amount of numerical integration required.
}
\author{
John T. Haman
}
