% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{asymmetric_laplace_dist}
\alias{asymmetric_laplace_dist}
\alias{alaplace_dist}
\title{Create a asymmetric Laplace distribution object for energy testing}
\usage{
asymmetric_laplace_dist(location = NULL, scale = NULL, skew = NULL)

alaplace_dist(location = NULL, scale = NULL, skew = NULL)
}
\arguments{
\item{location}{NULL, or a location parameter}

\item{scale}{NULL, or a positive scale parameter}

\item{skew}{NULL, or a positive Skewness parameter. Skew = 1 corresponds to
a symmetric Laplace distribution (though note the difference between the PDF
in this description and the one in \code{\link[=laplace_dist]{laplace_dist()}}).}
}
\value{
S3 data object containing the fields. This all used internally by
\code{\link[=energyfit]{energyfit()}} and \code{\link[=energyfit.test]{energyfit.test()}}.
\itemize{
\item name: String
\item composite_p: TRUE if test is composite
\item par: Distribution parameters, list of the formals.
\item sampler_par: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item par_domain: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item support: Function to check that data \code{x} can be tested against \code{dist}
\item sampler: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item EYY: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item EXYhat: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in ef.test or ef).
\item xform: Function that may be used to transform x. Only available in certain
distribution objects.
\item statistic: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item notes: Distribution specific messages.
}
}
\description{
This is exactly the distribution corresponding to the pdf

\deqn{
  f(x | \theta, \sigma, \kappa) =
  \frac{\sqrt{2}\kappa}{\sigma(1 + \kappa^2)}
  \begin{cases}
    \exp\Big( -\frac{\sqrt{2} \kappa |x - \theta|}{\sigma} \Big),
    & x \ge \theta, \\[6pt]
    \exp\Big( -\frac{\sqrt{2} |x - \theta|}{\kappa \sigma} \Big),
    & x < \theta.
  \end{cases}
, }

where \eqn{\theta} = \code{location}, \eqn{\sigma} = \code{scale}, and \eqn{\kappa} = \code{skew}.
}
\examples{

d <- asymmetric_laplace_dist(0, 1, .5)
x <- d$sampler(10, d$par)

ef(x, d)


}
\author{
John T. Haman
}
