% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{energyfit}
\alias{energyfit}
\alias{ef}
\title{Goodness-of-fit tests via the energy of data}
\usage{
energyfit(x, dist, nsim = 100)
}
\arguments{
\item{x}{A numeric vector.}

\item{dist}{A object of class GOFDist. The distribution to test \code{x} against.
GOFDist objects are created with the various \code{name_dist()} functions in
this package. See, for example, \code{\link[=normal_dist]{normal_dist()}} for details on these class
objects .}

\item{nsim}{A positive integer. The number of parametric bootstrap
replicates taken to calculate the p-value.}
}
\value{
An object of class `htest' representing the result of the energy
goodness-of-fit hypothesis test. The htest object has the list elements:
\itemize{
\item method: Simple or Composite
\item data.name
\item distribution: The distribution object created to test
\item parameter: List of parameters if the test is simple
\item nsim: Number of bootstrap replicates
\item composite_p: TRUE/FALSE composite predicate
\item statistic: The value of the energy statistic (\eqn{Q=nE^*})
\item p.value
\item sim_reps: bootstrap simulation of energy statistic
\item estimate: Any parameter estimates, if the test is composite
}
}
\description{
This is an alternative interface that provides the same test as
\code{\link[=energyfit.test]{energyfit.test()}}, but allows the user to pass a distribution object like
\code{\link[=normal_dist]{normal_dist()}} (Distribution objects are specific to this package). The
advantage is that you do not need to pass distribution parameters into a
\code{...} argument as in \code{energyfit.test}. \code{energyfit.test} uses this function
under the hood, but it's perfectly suitable for the user to use as well.
}
\details{
There are two types of goodness-of-fit tests covered by the energyfit
function, \emph{simple} and \emph{composite}. It's important to know the difference to
use this package, because yield different results. Simple GOF tests test the
data \code{x} against a specific distribution with \emph{known parameters} that you
must pass to energyfit.test in the ellipsis agrument (...). You should use a
simple GOF test if you wish to test questions like
"my data is Normal with mean 1 and sd 2". energyfit.test can also conduct
\emph{some} composite GOF tests. A composite test is performed if no parameters
are passed in the ellpisis argument (...). You should conduct a composite
test if your research question is
"my data is Normal, but I don't know what the parameters are." Obviously,
this composite question is much more common in practice.

All the composite tests in energyfit assume that none of the parameters are
known. So while there is a statistical test of Normality with known mean and
unknown sd, this is not implemented in the energygof package. So, either
pass all the distribution parameters or none of them. (In the special case
of the Normal distribution, you can use the energy package to test the GOF
hypothesis with any combination of known and known parameters.)
#'
For each test, energyfit.test calculates the test statistic and a p-value.
In all cases the p-value is calculated via parametric bootstrap. For large \code{nsim},
the p-values should be reasonably honest in small-ish samples.You should set
nsim to be a very large number in practice. I recommend at least 10,000. The
default value (100) is not a robust choice.
}
\section{About Energy}{


Székely, G. J., & Rizzo, M. L. (2023) provide the motivation:

" Data energy is a real number (typically a nonnegative number) that depends
on the distances between data. This concept is based on the notion of
Newton’s gravitational potential energy, which is also a function of the
distance between bodies. The idea of data energy or energy statistics is to
consider statistical observations (data) as heavenly bodies governed by the
potential energy of data, which is zero if and only if an underlying
statistical hypothesis is true."

The notation X' indicates that X' is an independent and identically
distributed copy of X.

If X and Y are independent and E(|X|^s + |Y|^s) is finite, then for 0 < s <
2,

\deqn{2E|X-Y|^s - E|X-X'|^s - E|Y-Y'|^s \ge 0.}

Equality is attained if and only if X and Y are identically distributed. The
left side of the equation is the energy between X and Y. Energy can be
generalized to multivariate data and even more exotic data types, but in
this R package, we only treat univariate data.

The concept of data energy between two random variables can be adapted to
the one sample goodness of fit problem. The one-sample s-energy is

\deqn{E^* = \frac{2}{n} \sum_i E|x_i - Y|^s - E|Y-Y'|^s - \frac{1}{n^2}
\sum_i \sum_j |x_i - x_j|^s,}

when \eqn{0 < s < 2} and \eqn{E|X|^s, E|Y|^s < \infty.}

In most tests in the energyfit package s = 1. In some cases (Pareto, Cauchy,
Stable), E|Y| is not finite, so we need to use an s < 1. This is done by
passing \code{pow} into \code{...} (but in all tests a default \code{pow} is provided)

In the one-sample goodness-of-fit regime, we wonder if \eqn{x_i ~ X} (where
the distribution of X is hidden) follows the same distribution as Y, which
is specified. If X and Y have the same distribution, then \eqn{Q = nE^*} is
a quadratic form of centered Gaussian random variables with expected value
\eqn{E|Y-Y'|^s}. If X and Y differ, then Q goes to Inf. So, Q provides a
consistent goodness-of-fit test, even in some situations where E|Y| is not
finite. And that's what energyfit.test does. Asymptotic theory of
V-statistics can be applied to prove that tests based on Q are statistically
consistent goodness-of-fit tests.
}

\section{Distributions available to test}{


The table summarizes the available distributions, their parameters, and
whether a composite GOF test may be performed.

\tabular{llll}{
\strong{Distribution} \tab \strong{Function} \tab \strong{Paramater} \tab \strong{CompositeAllowed} \cr
\tab                        \tab                                       \tab    \cr
Normal             \tab normal_dist            \tab mean, sd                              \tab Yes\cr
Normal             \tab normal_dist            \tab mean, sd                              \tab Yes\cr
Uniform            \tab uniform_dist           \tab min, max                              \tab No \cr
Exponential        \tab exponential_dist       \tab rate                                  \tab Yes\cr
Poisson            \tab poisson_dist           \tab lambda                                \tab Yes\cr
Bernoulli          \tab bernoulli_dist         \tab prob                                  \tab No \cr
Binomial           \tab binomial_dist          \tab prob                                  \tab Yes\cr
Beta               \tab beta_dist              \tab shape1, shape2                        \tab No \cr
Half-Normal        \tab halfnormal_dist        \tab theta                                 \tab No \cr
Laplace            \tab laplace_dist           \tab location, scale                       \tab No \cr
Log-normal         \tab lognormal_dist         \tab meanlog, sdlog                        \tab No \cr
Asymmetric Laplace \tab asymmetriclaplace_dist \tab location, scale, skew                 \tab No \cr
Weibull            \tab weibull_dist           \tab shape, scale                          \tab No \cr
Gamma              \tab gamma_dist             \tab shape, rate                           \tab No \cr
Chi Squared        \tab chisq_dist             \tab df                                    \tab No \cr
Inverse Gaussion   \tab inversegaussian_dist   \tab mean, shape                            \tab No \cr
Pareto             \tab pareto_dist            \tab scale, shape, pow, r                  \tab No \cr
Cauchy             \tab cauchy_dist            \tab location, scale, pow                  \tab No \cr
Stable             \tab stable_dist            \tab location, scale, skew, stability, pow \tab No
}
}

\examples{
## Simple, normal test
energyfit(rnorm(10), normal_dist(0, 1), nsim = 10)

## Simple, Poisson test
energyfit(rpois(10,1), poisson_dist(1), nsim = 0) # No p-value

## Composite Normal test
energyfit(rnorm(10), normal_dist(), nsim = 10)

## In real application, you should increase nsim to at least 10,000.


}
\references{
Székely, G. J., & Rizzo, M. L. (2023). The energy of data and distance
correlation. Chapman and Hall/CRC.

Székely, G. J., & Rizzo, M. L. (2013). Energy statistics: A class of
statistics based on distances. Journal of statistical planning and
inference, 143(8), 1249-1272.

Li, Y. (2015). Goodness-of-fit tests for Dirichlet distributions with
applications. Bowling Green State University.

Rizzo, M. L. (2002). A new rotation invariant goodness-of-fit test (PhD
thesis). Bowling Green State University

Haman, J. T. (2018). The energy goodness-of-fit test and EM type estimator
for asymmetric Laplace distributions (Doctoral dissertation, Bowling Green
State University).

Ofosuhene, P. (2020). The energy goodness-of-fit test for the inverse
Gaussian distribution (Doctoral dissertation, Bowling Green State
University).

Rizzo, M. L. (2009). New goodness-of-fit tests for Pareto distributions.
ASTIN Bulletin: The Journal of the IAA, 39(2), 691-715.
}
\seealso{
\link[stats]{Distributions} for a list of distributions available
in most R installations. \code{\link[energy:normalGOF]{energy::normal.test()}} for the energy
goodness-of-fit test with unknown parameters. See \code{\link[energy:poisson]{energy::poisson.mtest()}}
for a different poisson goodness-of-fit test based on mean distances. The
tests for (multivariate) Normal in the energy package are implemented
with compiled code, and are faster than the one available in the energyfit
package.
}
\author{
John T. Haman
}
