% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{energyfit.test}
\alias{energyfit.test}
\alias{ef.test}
\title{Goodness-of-fit tests via the energy of data}
\usage{
\method{energyfit}{test}(
  x,
  dist = c("uniform", "exponential", "bernoulli", "binomial", "geometric", "normal",
    "gaussian", "beta", "poisson", "lognormal", "lnorm", "laplace", "doubleexponential",
    "asymmetriclaplace", "alaplace", "inversegaussian", "halfnormal", "chisq",
    "chisquared", "gamma", "weibull", "cauchy", "stable", "pareto"),
  nsim = 100,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{dist}{A string. The distribution to test \code{x} against.}

\item{nsim}{A positive integer. The number of parametric bootstrap
replicates taken to calculate the p-value.}

\item{...}{Parameters of the distribution \code{dist}. For distributions in the R
\verb{stats' library, parameter argument names are identical. To test the _composite_ goodness-of-fit hypothesis that }x\verb{is distributed according to the _family of distributions_}dist\verb{, don't pass parameters in }...\verb{. For generalized energy tests, you can also optionally pass }pow` here.}
}
\value{
An object of class `htest' representing the result of the energy
goodness-of-fit hypothesis test. The htest object has the list elements:
\itemize{
\item method: Simple or Composite
\item data.name
\item distribution: The distribution object created to test
\item parameter: List of parameters if the test is simple
\item nsim: Number of bootstrap replicates
\item composite_p: TRUE/FALSE composite predicate
\item statistic: The value of the energy statistic (\eqn{Q=nE^*})
\item p.value
\item sim_reps: bootstrap simulation data
\item estimate: Any parameter estimates, if the test is composite
}
}
\description{
Perform a goodness-of-fit test of univariate data \code{x} against a
distribution \code{dist}. Both simple (known parameter) and composite (unknown
parameter) cases are supported, but not all distributions allow for a
composite test. P-values are determined via parametric bootstrap. For
distributions where E|Y| is not finite (Cauchy, Pareto, Stable), a
generalized energy goodness-of-fit test is performed, and an additional
tuning parameter \code{pow} is required.
}
\details{
There are two types of goodness-of-fit tests covered by the energyfit
function, \emph{simple} and \emph{composite}. It's important to know the difference to
use this package, because yield different results. Simple GOF tests test the
data \code{x} against a specific distribution with \emph{known parameters} that you
must pass to energyfit.test in the ellipsis agrument (...). You should use a
simple GOF test if you wish to test questions like
"my data is Normal with mean 1 and sd 2". energyfit.test can also conduct
\emph{some} composite GOF tests. A composite test is performed if no parameters
are passed in the ellpisis argument (...). You should conduct a composite
test if your research question is
"my data is Normal, but I don't know what the parameters are." Obviously,
this composite question is much more common in practice.

All the composite tests in energyfit assume that none of the parameters are
known. So while there is a statistical test of Normality with known mean and
unknown sd, this is not implemented in the energygof package. So, either
pass all the distribution parameters or none of them. (In the special case
of the Normal distribution, you can use the energy package to test the GOF
hypothesis with any combination of known and known parameters.)
#'
For each test, energyfit.test calculates the test statistic and a p-value.
In all cases the p-value is calculated via parametric bootstrap. For large \code{nsim},
the p-values should be reasonably honest in small-ish samples.You should set
nsim to be a very large number in practice. I recommend at least 10,000. The
default value (100) is not a robust choice.
}
\section{Distributions-in-energyfit}{


The table summarizes the available distributions, their parameters, and
whether a composite GOF test may be performed.

\tabular{llll}{
\strong{Distribution} \tab \strong{Function} \tab \strong{Paramater} \tab \strong{CompositeAllowed} \cr
\tab                        \tab                                       \tab    \cr
Normal             \tab normal_dist            \tab mean, sd                              \tab Yes\cr
Uniform            \tab uniform_dist           \tab min, max                              \tab No \cr
Exponential        \tab exponential_dist       \tab rate                                  \tab Yes\cr
Poisson            \tab poisson_dist           \tab lambda                                \tab Yes\cr
Bernoulli          \tab bernoulli_dist         \tab prob                                  \tab No \cr
Binomial           \tab binomial_dist          \tab prob                                  \tab Yes\cr
Beta               \tab beta_dist              \tab shape1, shape2                        \tab No \cr
Half-Normal        \tab halfnormal_dist        \tab theta                                 \tab No \cr
Laplace            \tab laplace_dist           \tab location, scale                       \tab No \cr
Log-normal         \tab lognormal_dist         \tab meanlog, sdlog                        \tab No \cr
Asymmetric Laplace \tab asymmetriclaplace_dist \tab location, scale, skew                 \tab No \cr
Weibull            \tab weibull_dist           \tab shape, scale                          \tab No \cr
Gamma              \tab gamma_dist             \tab shape, rate                           \tab No \cr
Chi Squared        \tab chisq_dist             \tab df                                    \tab No \cr
Inverse Gaussion   \tab inversegaussian_dist   \tab mu, lambda                            \tab No \cr
Pareto             \tab pareto_dist            \tab scale, shape, pow, r                  \tab No \cr
Cauchy             \tab cauchy_dist            \tab location, scale, pow                  \tab No \cr
Stable             \tab stable_dist            \tab location, scale, skew, stability, pow \tab No
}
}

\section{About-Energy}{


Székely, G. J., & Rizzo, M. L. (2023) provide the motivate:

Data energy is a real number (typically a nonnegative number) that depends
on the distances between data. This concept is based on the notion of
Newton’s gravitational potential energy, which is also a function of the
distance between bodies. The idea of data energy or energy statistics is to
consider statistical observations (data) as heavenly bodies governed by the
potential energy of data, which is zero if and only if an underlying
statistical hypothesis is true.

The concept of data energy between two random variables can be adapted to
the one sample goodness of fit problem. The one-sample s-energy is

\deqn{E^* = \frac{2}{n} \sum_i E|x_i - Y|^s - E|Y-Y'|^s - \frac{1}{n^2}
\sum_i \sum_j |x_i - x_j|^s,}

when \eqn{0 < s < 2} and \eqn{E|X|^s, E|Y|^s < \infty.}

In most tests in the energyfit package s = 1. In some cases (Pareto, Cauchy,
Stable), E|Y| is not finite, so we need to use an s < 1. This is done by
passing \code{paw} into \code{...}. (See examples)

In the one-sample goodness-of-fit regime, we wonder if \eqn{x_i ~ X} (which is
hidden) follows the same distribution as Y, which is specified. If X and Y
have the same distribution, then \eqn{Q = nE^*} is a quadratic form of
centered Gaussian random variables with expected value \eqn{E|Y-Y'|}. If X and Y
differ, then Q goes to Inf. So, Q can be used to test goodness-of-fit, even
in some situations where E|Y| is not finite. And that's what energyfit.test
does.
}

\examples{
x <- rnorm(10)

## Composite energy goodness-of-fit test (test for Normality with unknown
## parameters)

energyfit.test(x, "normal", nsim = 10)

## Simple energy goodness-of-fit test (test for Normality with known
## parameters). ef is an alias for energyfit.test.

ef.test(x, "normal", nsim = 10, mean = 0, sd = 1)

## Alternatively, use the energyfit generic directly so that you do not need
## to pass parameter names into `...`

energyfit(x, normal_dist(0, 1), nsim = 10)

## Simple energy goodness-of-fit test for Weibull distribution

y <- rweibull(10, 1, 1)
energyfit.test(y, "weibull", shape = 1, scale = 3, nsim = 10)

## Alternatively, use the energyfit generic directly, which is slightly less
## verbose. ef is an alias for energyfit.

ef(y, weibull_dist(1, 3), nsim = 10

## energyfit does not support "partially composite" GOF tests, so this will
## result in an error.

## Conduct a generalized GOF test. Pow is the exponent s in the generalized
#energy statistic. Pow is only necessary when testing Stable, Cauchy, and
#Pareto distributions. If you don't set a pow, there is a default for each
#of the distributions, but the default isn't necessarily better than any
#other number.

x <- rcauchy(10)
ef(x, cauchy_dist(location = 0, scale = 1, pow = 0.5))

\dontrun{
energyfit.test(x, "normal", mean = 0, nsim = 10) # sd is missing
}

}
\references{
Székely, G. J., & Rizzo, M. L. (2023). The energy of data and distance
correlation. Chapman and Hall/CRC.

Székely, G. J., & Rizzo, M. L. (2013). Energy statistics: A class of
statistics based on distances. Journal of statistical planning and
inference, 143(8), 1249-1272.

Li, Y. (2015). Goodness-of-fit tests for Dirichlet distributions with
applications. Bowling Green State University.

Rizzo, M. L. (2002). A new rotation invariant goodness-of-fit test (PhD
thesis). Bowling Green State University

Haman, J. T. (2018). The energy goodness-of-fit test and EM type estimator
for asymmetric Laplace distributions (Doctoral dissertation, Bowling Green
State University).

Ofosuhene, P. (2020). The energy goodness-of-fit test for the inverse
Gaussian distribution (Doctoral dissertation, Bowling Green State
University).

Rizzo, M. L. (2009). New goodness-of-fit tests for Pareto distributions.
ASTIN Bulletin: The Journal of the IAA, 39(2), 691-715.
}
\seealso{
\link[stats]{Distributions} for a list of distributions available
in most R installations. \link[energy:normalGOF]{energy::normal.test} for the energy
goodness-of-fit test with unknown parameters. See \link[energy:poisson]{energy::poisson.mtest}
for a different poisson goodness-of-fit test based on mean distances. The
tests for (multivariate) Normal in the \link{energy} package are implemented
with compiled code, and are faster than the one available in the energyfit
package.
}
\author{
John T. Haman
}
