% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{lognormal_dist}
\alias{lognormal_dist}
\title{Create a log-normal distribution object for energy testing}
\usage{
lognormal_dist(meanlog = NULL, sdlog = NULL)
}
\arguments{
\item{meanlog}{NULL or as in \code{\link[=rlnorm]{rlnorm()}}, must be length 1.}

\item{sdlog}{NULL or as in \code{\link[=rlnorm]{rlnorm()}}, must be length 1.

d <- lognormal_dist(0, 1)
x <- d$sampler(10, d$par)

eg(x, d)}
}
\value{
S3 data object containing the fields. This all used internally by
\code{\link[=energyGOF]{energyGOF()}} and \code{\link[=energyGOF.test]{energyGOF.test()}}.
\itemize{
\item name: String
\item composite_p: TRUE if test is composite
\item par: Distribution parameters, list of the formals.
\item sampler_par: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item par_domain: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item support: Function to check that data \code{x} can be tested against \code{dist}
\item sampler: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item EYY: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item EXYhat: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in eg.test or eg).
\item xform: Function that may be used to transform x. Only available in certain
distribution objects.
\item statistic: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item notes: Distribution specific messages.
}
}
\description{
Create an S3 object that sets all the required data needed by
energyGOF to execute the energy goodness-of-fit test against a log-normal
distribution. If \code{meanlog} and \code{sdlog} are both \code{NULL}, a composite test is
performed.
}
\author{
John T. Haman
}
