% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{normal_dist}
\alias{normal_dist}
\title{Create a Normal distribution object for energy testing}
\usage{
normal_dist(mean = NULL, sd = NULL)
}
\arguments{
\item{mean}{NULL, or if specified, same as \code{\link[=rnorm]{rnorm()}}, but must be length 1.}

\item{sd}{NULL, or if specified, Same as \code{\link[=rnorm]{rnorm()}}, but must be length 1}
}
\value{
S3 data object containing the fields. This all used internally by
\code{\link[=energyfit]{energyfit()}} and \code{\link[=energyfit.test]{energyfit.test()}}.
\itemize{
\item name: String
\item composite_p: TRUE if test is composite
\item par: Distribution parameters, list of the formals.
\item sampler_par: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item par_domain: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item support: Function to check that data \code{x} can be tested against \code{dist}
\item sampler: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item EYY: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item EXYhat: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in ef.test or ef).
\item xform: Function that may be used to transform x. Only available in certain
distribution objects.
\item statistic: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item notes: Distribution specific messages.
}
}
\description{
Create an S3 object that sets all the required data needed by
energyfit to execute the eneregy goodness-of-fit test against a normal
distribution. If \code{mean} and \code{sd} are both NULL, perform a composite test.
}
\author{
John T. Haman
}
