% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyGOF.R
\name{pareto_dist}
\alias{pareto_dist}
\title{Create a Pareto (type I) distribution object for energy testing}
\usage{
pareto_dist(scale = NULL, shape = NULL, pow = shape/4)
}
\arguments{
\item{scale}{NULL or a positive scale parameter}

\item{shape}{NULL or a positive shape parameter. If shape > 1, shape is used
to transform x}

\item{pow}{Optional exponent of the energy test. Pow must be less than
shape/2. If shape > 1 and pow != 1, pow will be scaled to pow/(4*shape) to
ensure pow < 1/2}
}
\value{
S3 data object containing following the fields.
\itemize{
\item \code{name}: String
\item \code{composite_p}: Composite predicate. TRUE if test is composite.
\item \code{par}: Distribution parameters, list of the formals.
\item \code{sampler_par}: Distribution parameters used for the calculation of energy
statistic. These may be different than \code{par}.
\item \code{par_domain}: Function used to ensure \code{par} and \code{sampler_par} are valid for
this distribution
\item \code{support}: Function to check that data \code{x} can be tested against \code{dist}
\item \code{sampler}: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item \code{EYY}: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the
generalized test.)
\item \code{EXYhat}: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or
\eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according
to \code{dist} and x is the data under test (which is passed in \code{eg.test} or \code{eg}).
\item \code{xform}: Function that may be used to transform x. Only available in certain
distribution objects.
\item \code{statistic}: Function that returns a list of maximum likelihood estimates.
Only available in certain distribution objects.
\item \code{notes}: Distribution specific messages. Only used in certain distribution
objects.
}
}
\description{
Create an S3 object that sets all the required data needed by
energyGOF to execute the energy goodness-of-fit test against a normal
distribution. If \code{mean} and \code{sd} are both NULL, perform a composite test.
}
\details{
If shape > 1, the energy test is more difficult, so X is
transformed to X^shape ~ Pareto(scale^shape, 1). If both shape and scale
are NULL, then a composite test is performed.
}
\examples{
d <- pareto_dist(1, .5)
x <- d$sampler(10, d$par)
eg(x, d)

}
\author{
John T. Haman
}
