% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyfit.R
\name{pareto_dist}
\alias{pareto_dist}
\title{Create a Pareto (type I) distribution object for energy testing}
\usage{
pareto_dist(scale = NULL, shape = NULL, pow = shape/2, r = shape)
}
\arguments{
\item{scale}{Positive scale parameter}

\item{shape}{Positive shape parameter. If shape > 1, r is used to transform x}

\item{pow}{exponent of the energy test. Pow must be less than shape.}

\item{r}{Used to transform x in case shape > 1 and pow non-zero}

\item{skew}{Skewness parameter}
}
\value{
S3 data object containing the fields. This all used internally by \code{\link[=energyfit]{energyfit()}} and \code{\link[=energyfit.test]{energyfit.test()}}.
\itemize{
\item name: String
\item composite_p: TRUE if test is composite
\item par: Distribution parameters, list of the formals.
\item sampler_par: Distribution parameters used for the calculation of energy statistic. These may be different than \code{par}.
\item par_domain: Function used to ensure \code{par} and \code{sampler_par} are valid for this distribution
\item support: Function to check that data \code{x} can be tested against \code{dist}
\item sampler: Function used for rng by \code{\link[boot:boot]{boot::boot()}}
\item EYY: Function to compute \eqn{E|Y-Y'|} (or \eqn{E|Y-Y'|^{pow}}, for the generalized test.)
\item EXYhat: Function to compute \eqn{\frac{1}{n} \sum_i E|x_i - Y|} (or  \eqn{\frac{1}{n} \sum_i E|x_i - Y|^{pow}}), where Y is distributed according to dist and x is the data under test (which is passed in ef.test or ef).
\item xform: Function that may be used to transform x. Only available in certain distribution objects.
\item statistic: Function that returns a list of maximum likelihood estimates. Only available in certain distribution objects.
}
}
\description{
If shape > 1, the energy test is more difficult, so X is transformed to X^r ~ Pareto(scale*shape, shape/r), so that shape/r <= 1, and pow < shape/r.
}
\author{
John T. Haman
}
